Fixing the autoloader warnings and moving namespaces
