+++ packages/article/tests/Unit/Infrastructure/Sulu/Content/ResourceLoader/ArticleResourceLoaderTest.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\ArticleBundle\Tests\Unit\Infrastructure\Sulu\Content\ResourceLoader;
+namespace Sulu\Article\Tests\Unit\Infrastructure\Sulu\Content\ResourceLoader;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
