+++ packages/content/tests/Unit/Infrastructure/Symfony/HttpKernel/Compiler/SeoFormPassTest.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Content\Tests\Unit\Content\Infrastructure\Symfony\HttpKernel\Compiler;
+namespace Sulu\Content\Tests\Unit\Infrastructure\Symfony\HttpKernel\Compiler;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
