+++ packages/page/tests/Unit/Infrastructure/Sulu/Content/DataMapper/NavigationContextDataMapperTest.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\Page\Tests\Unit\Infrastructure\Sulu\Content\DataMapper;
+namespace Sulu\Page\Tests\Unit\Infrastructure\Sulu\Content\DataMapper;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
