+++ packages/page/tests/Unit/Infrastructure/Sulu/Content/ResourceLoader/PageResourceLoaderTest.php
@@ -9,7 +9,7 @@
  * with this source code in the file LICENSE.
  */
 
-namespace Sulu\Bundle\Page\Tests\Unit\Infrastructure\Sulu\Content\ResourceLoader;
+namespace Sulu\Page\Tests\Unit\Infrastructure\Sulu\Content\ResourceLoader;
 
 use PHPUnit\Framework\TestCase;
 use Prophecy\Argument;
